Fix "multiple values for argument 'foreign_pre_chain'" issue
Not sure how that ever worked. Possibly a change in a recent structlog
update.

The actual issue is that a superfluous `self` was passed in the `super`
call. The lead to `foreign_pre_chain` being filled positionally as well
as from the kwargs.

Since `__init__` isn't doing anything anyway we can just remove it.