Add exit codes to SP run routine
Introduces the following Exit code ranges:

    Exit code 1x
    There was a problem when starting up the SP, nodes, deploying tokens
    or setting up services. This points at an issue in the SP and of of its
    components.

    Exit code 2x
    There was an error when parsing or evaluating the given scenario
    definition file. This may be a syntax- or logic-related issue.

    Exit code 3x
    There was an assertion error while executing the scenario. This points
    to an error in a `raiden` component (the client, services or contracts).

The default assumed exit code is 1, and it's raised when
an exception was unhandled. If the raised exception
IS handled, the above exit code will be applied.

It's favorable to be more granular with exit codess in the future,
which is why it's advised to integrate an exit code value
to our custom exceptions, which we can then read from
the command functions. For now, the code values are assigned
at the top level of the cli command function.