Make sure that exceptions are visible in the UI
The UI captures all log output, so we must not quit before the user has
seen the exception that causes us to exit.

Using `.ready()` instead `dead` is necessary to make sure that we show
the UI at all, since `dead` is true before the UI thread was run once.