Avoid inconsistent RECLAIM_MIN_BALANCE
Before this commit, `RECLAIM_MIN_BALANCE` was not the same as the tx
cost for the reclaim. If it's bigger, we would not reclaim some small
funds. If it's smaller, we would get an error because we would try to
reclaim a negative value.

The obvious fix is to use the tx cost as the minimum reclaim balance.