Add reclamation for UDC deposits and ERC20 tokens
This is unnecessary when we use mintable tokens, but when testing with
real tokes on mainnet, this recovers valuable tokens. Without the new
cli parameters, the reclaim script has the old behavior.

This implementation is quite slow and could be sped up by creating
transactions in parallel. But reclaiming on mainnet should be rare, so
the current simple approach should be sufficient.

Closes https://github.com/raiden-network/scenario-player/issues/543