Fixed problems with address order
There is no guarantee on the order of the address used in the channel
open event, therefore the code must check if the current address is in
position 1 or 2, and use the other as the partner address. Previously
the code was assuming that the current address was always on position 1.