Include environment files in package
This makes then available wherever the package is installed. I added a
symlink from the old location for backwards compatibility.

Flit automatically includes all files in the package in the release,
thus the lack of a `MANIFEST.in`. I could not find a way to include the
environment without moving them into the package.