Fix nonce error when snapshots are present
Having multiple node_dirs for the same account caused nonce errors
because a JSONRPCClient was created for each node_dir. Keep all of those
in a shared address->client mapping to avoid this.

The problem could also have been avoided by deduplicating the
candidates, but this solution seems more robust with regards to future
code changes to me.