Add AssertChannelSettledEvent
Since the call-arguments for `participant1` and `participant2` are
unordered and the order is not enforced based on who calls the function,
both possibilities have to be checked when making assertions on
ChannelSettled events that are coop-settle related.
This commit introduces the `AssertChannelSettledEvent` task,
that allows for more advanced assertions related to ChannelSettled
events.